# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tic Tac Toe game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomRequest {
  hostName: string;
  /**
   * @minimum 3
   * @maximum 10
   */
  boardSize?: number;
  /** @minimum 3 */
  winLength?: number;
  /** Seconds per turn, null for unlimited */
  timeLimit?: number | null;
  symbolX?: string;
  symbolO?: string;
  theme?: string;
}

export type RoomResponseStatus =
  (typeof RoomResponseStatus)[keyof typeof RoomResponseStatus];

export const RoomResponseStatus = {
  waiting: "waiting",
  playing: "playing",
  finished: "finished",
} as const;

export type RoomResponseCurrentTurn =
  (typeof RoomResponseCurrentTurn)[keyof typeof RoomResponseCurrentTurn];

export const RoomResponseCurrentTurn = {
  X: "X",
  O: "O",
} as const;

export interface RoomResponse {
  id: string;
  roomCode: string;
  hostName: string;
  guestName?: string | null;
  boardSize: number;
  winLength: number;
  timeLimit?: number | null;
  symbolX: string;
  symbolO: string;
  theme: string;
  status: RoomResponseStatus;
  board: (string | null)[];
  currentTurn: RoomResponseCurrentTurn;
  winner?: string | null;
  createdAt: string;
}

export interface LeaderboardEntry {
  id: string;
  playerName: string;
  wins: number;
  losses: number;
  draws: number;
  totalGames: number;
  winRate: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetLeaderboardParams = {
  limit?: number;
};