# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tic Tac Toe game API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomRequest {
  hostName: string;
  /**
   * @minimum 3
   * @maximum 10
   */
  boardSize?: number;
  /** @minimum 3 */
  winLength?: number;
  /** Seconds per turn, null for unlimited */
  timeLimit?: number | null;
  symbolX?: string;
  symbolO?: string;
  theme?: string;
}