# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tic Tac Toe game API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  id: string;
  playerName: string;
  wins: number;
  losses: number;
  draws: number;
  totalGames: number;
  winRate: number;
}