# lib/api-zod/src/generated/types/roomResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tic Tac Toe game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomResponseCurrentTurn } from "./roomResponseCurrentTurn";
import type { RoomResponseStatus } from "./roomResponseStatus";

export interface RoomResponse {
  id: string;
  roomCode: string;
  hostName: string;
  guestName?: string | null;
  boardSize: number;
  winLength: number;
  timeLimit?: number | null;
  symbolX: string;
  symbolO: string;
  theme: string;
  status: RoomResponseStatus;
  board: (string | null)[];
  currentTurn: RoomResponseCurrentTurn;
  winner?: string | null;
  createdAt: Date;
}